Fix Qwen3.5 MoE models using dense MLP instead of SparseMoeBlock
The base qwen3_5 DecoderLayer unconditionally created a dense Qwen3_5MLP,
ignoring num_experts from the model config. MoE models (35B-A3B, 122B-A10B)
loaded through this path compute the wrong function and produce garbage.

Add MoE config fields (num_experts, etc.) to TextConfig with safe zero
defaults, add Qwen3_5SparseMoeBlock to the base module, and make
DecoderLayer select SparseMoeBlock when num_experts > 0. This mirrors
the pattern in mlx_lm/models/qwen3_5.py. The qwen3_5_moe module is
simplified to thin wrappers over the now-unified base.

Fixes #1073
